rename NAMESPACES to NAMESPACE, and move city from user to data state
