+++ src/reducer/data/data.js
@@ -2,11 +2,13 @@ import ModelOffer from '../../model-offer';
 
 
 const initialState = {
+  city: `No cities`,
   listOffers: [],
 };
 
 const ActionType = {
   ADD_LIST_OFFERS: `ADD_LIST_OFFERS`,
+  CHANGE_CITY: `CHANGE_CITY`,
   LOAD_OFFERS: `LOAD_OFFERS`,
 };
 
@@ -15,6 +17,12 @@ const ActionCreator = {
     type: ActionType.ADD_LIST_OFFERS,
     payload: list,
   }),
+
+  changeCity: (newCity) => ({
+    type: ActionType.CHANGE_CITY,
+    payload: newCity,
+  }),
+
 
   loadOffers: (offers) => {
     return {
@@ -46,6 +54,11 @@ const reducer = (state = initialState, action) =>{
     case ActionType.ADD_LIST_OFFERS:
       return Object.assign({}, state, {
         listOffers: action.payload,
+      });
+
+    case ActionType.CHANGE_CITY:
+      return Object.assign({}, state, {
+        city: action.payload,
       });
 
     case ActionType.LOAD_OFFERS:
