+++ src/reducer/user/user.js
@@ -1,5 +1,4 @@
 const initialState = {
-  city: `No cities`,
   isAuthorizationRequired: false,
   id: 1,
   email: `Oliver.conner@gmail.com`,
@@ -9,18 +8,12 @@ const initialState = {
 };
 
 const ActionType = {
-  CHANGE_CITY: `CHANGE_CITY`,
   REQUIRED_AUTHORIZATION: `REQUIRED_AUTHORIZATION`,
   ADD_USER_DATA: `ADD_USER_DATA`,
   LOG_IN: `LOG_IN`,
 };
 
 const ActionCreator = {
-  changeCity: (newCity) => ({
-    type: ActionType.CHANGE_CITY,
-    payload: newCity,
-  }),
-
   requireAuthorization: (status) => {
     return {
       type: ActionType.REQUIRED_AUTHORIZATION,
@@ -66,11 +59,6 @@ const Operation = {
 
 const reducer = (state = initialState, action) =>{
   switch (action.type) {
-    case ActionType.CHANGE_CITY:
-      return Object.assign({}, state, {
-        city: action.payload,
-      });
-
     case ActionType.REQUIRED_AUTHORIZATION:
       return Object.assign({}, state, {
         isAuthorizationRequired: action.payload,
